+++ backend/src/main/java/com/zapiszto/controllers/exercises/dto/NewExerciseTrainingDto.java
@@ -21,7 +21,7 @@ public class NewExerciseTrainingDto {
   UUID dictExerciseId;
   int quantity;
   UUID dictQuantityTypeId;
-  int volume;
+  Float volume;
   UUID dictUnitId;
   String notes;
 }
