+++ backend/src/main/java/com/zapiszto/controllers/exercises/dto/UpdateVolumeDto.java
@@ -15,5 +15,5 @@ import lombok.experimental.FieldDefaults;
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class UpdateVolumeDto {
-  Integer volume;
+  Float volume;
 }
